functions/requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-ai-vision-imageanalysis
azure-cosmos
